requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-login>=0.6.0
werkzeug>=2.3.0
python-dateutil>=2.8.0

mysql-connector-python>=8.0.0
